+++ src/components/themed/ThemedScene.tsx
@@ -11,6 +11,8 @@ import { cacheStyles } from 'react-native-patina'
 import FontAwesome5 from 'react-native-vector-icons/FontAwesome5'
 
 import s from '../../common/locales/strings'
+import { useHandler } from '../../hooks/useHandler'
+import { Branding } from '../../types/Branding'
 import { fixSides, mapSides, sidesToPadding } from '../../util/sides'
 import { Theme, useTheme } from '../services/ThemeContext'
 import { DividerLine } from './DividerLine'
@@ -19,6 +21,8 @@ interface Props {
   children?: React.ReactNode
 
   // Header:
+  backButtonText?: string
+  branding?: Branding
   noUnderline?: boolean
   onBack?: () => void
   onSkip?: () => void
@@ -30,6 +34,8 @@ interface Props {
 
 export function ThemedScene(props: Props) {
   const {
+    backButtonText,
+    branding = {},
     children,
     onBack,
     onSkip,
@@ -37,10 +43,16 @@ export function ThemedScene(props: Props) {
     title,
     noUnderline = false
   } = props
+  const { parentButton } = branding
   const theme = useTheme()
   const styles = getStyles(theme)
 
-  const hasHeader = onBack != null || onSkip != null || title != null
+  const hasHeader =
+    onBack != null || onSkip != null || title != null || parentButton != null
+
+  const handleParentButtonPress = useHandler(() => {
+    if (parentButton != null) parentButton.callback()
+  })
 
   const containerStyle = {
     flex: 1,
@@ -58,17 +70,30 @@ export function ThemedScene(props: Props) {
         {!hasHeader ? null : (
           <View style={styles.headerButtons}>
             {onBack == null ? null : (
-              <TouchableOpacity style={styles.backButton} onPress={onBack}>
+              <TouchableOpacity style={styles.leftButton} onPress={onBack}>
                 <FontAwesome5
                   name="chevron-left"
                   size={theme.rem(1)}
-                  color={theme.primaryText}
+                  style={styles.buttonIcon}
                 />
+                {backButtonText == null ? null : (
+                  <Text style={styles.buttonText}>{backButtonText}</Text>
+                )}
               </TouchableOpacity>
             )}
             {onSkip == null ? null : (
-              <TouchableOpacity style={styles.skipButton} onPress={onSkip}>
-                <Text style={styles.skipText}>{s.strings.skip}</Text>
+              <TouchableOpacity style={styles.rightButton} onPress={onSkip}>
+                <Text style={styles.buttonText}>{s.strings.skip}</Text>
+              </TouchableOpacity>
+            )}
+            {parentButton == null || parentButton.text == null ? null : (
+              <TouchableOpacity
+                style={styles.rightButton}
+                onPress={handleParentButtonPress}
+              >
+                <Text style={parentButton.style || styles.buttonText}>
+                  {parentButton.text}
+                </Text>
               </TouchableOpacity>
             )}
           </View>
@@ -89,23 +114,30 @@ const getStyles = cacheStyles((theme: Theme) => ({
     flexDirection: 'row',
     height: theme.rem(3)
   },
-  backButton: {
+  leftButton: {
+    alignItems: 'center',
+    flexDirection: 'row',
     justifyContent: 'center',
     paddingHorizontal: theme.rem(1),
     position: 'absolute',
-    bottom: 0,
     left: 0,
     top: 0
   },
-  skipButton: {
+  rightButton: {
+    alignContent: 'center',
+    flexDirection: 'row',
     justifyContent: 'center',
     paddingHorizontal: theme.rem(1),
     position: 'absolute',
-    bottom: 0,
     right: 0,
     top: 0
   },
-  skipText: {
+  buttonIcon: {
+    color: theme.primaryText,
+    paddingRight: theme.rem(0.5),
+    height: theme.rem(1)
+  },
+  buttonText: {
     color: theme.primaryText,
     fontFamily: theme.fontFaceDefault,
     fontSize: theme.rem(1)